+++ component/transit/src/main/java/com/trimettransit/tracker/transit/TransitRepositoryImpl.kt
@@ -7,6 +7,7 @@ import com.trimettransit.tracker.model.Route
 import com.trimettransit.tracker.model.Stop
 import com.trimettransit.tracker.model.TripPoint
 import com.trimettransit.tracker.model.TripPlanResult
+import com.trimettransit.tracker.model.TripRequestOptions
 import com.trimettransit.tracker.model.TripRequestTime
 import com.trimettransit.tracker.model.VehiclePosition
 import com.trimettransit.tracker.model.repository.TransitRepository
@@ -72,8 +73,9 @@ class TransitRepositoryImpl(
     override suspend fun planTrip(
         from: TripPoint,
         to: TripPoint,
-        time: TripRequestTime
-    ): TripPlanResult? = TransitApi.fetchTripPlan(context, from, to, time)
+        time: TripRequestTime,
+        options: TripRequestOptions
+    ): TripPlanResult? = TransitApi.fetchTripPlan(context, from, to, time, options)
 }
 
 /**